Implemented DeleteEmailModal on list.component

+++ libs/mailroom/ngx/src/lib/modules/list/list.component.ts
@@ -6,6 +6,9 @@ import { Observable, Subject } from 'rxjs';
 import { EmailService } from '@tamu-gisc/mailroom/data-access';
 import { MailroomEmail } from '@tamu-gisc/mailroom/common';
 import { NotificationService } from '@tamu-gisc/common/ngx/ui/notification';
+import { ModalService } from '@tamu-gisc/ui-kits/ngx/layout/modal';
+
+import { DeleteEmailModalComponent } from './modal/delete-email-modal.component';
 
 @Component({
   selector: 'tamu-gisc-list',
@@ -16,7 +19,11 @@ export class ListComponent implements OnInit {
   private _$refresh: Subject<boolean> = new Subject();
   public $emails: Observable<Array<MailroomEmail>>;
 
-  constructor(public readonly emailService: EmailService, private ns: NotificationService) {}
+  constructor(
+    public readonly emailService: EmailService,
+    private ns: NotificationService,
+    private readonly modal: ModalService
+  ) {}
 
   public ngOnInit() {
     this.$emails = this._$refresh.pipe(
@@ -28,14 +35,24 @@ export class ListComponent implements OnInit {
   }
 
   public deleteEmail(email: MailroomEmail) {
-    this.emailService.deleteEmail(email.id).subscribe((removed) => {
-      if (removed) {
-        this.ns.preset('deleted_email_success');
+    this.modal
+      .open<{ deleteEmail: boolean }>(DeleteEmailModalComponent, {
+        data: {
+          message: email.id
+        }
+      })
+      .subscribe((proceedWithDelete) => {
+        if (proceedWithDelete) {
+          this.emailService.deleteEmail(email.id).subscribe((removed) => {
+            if (removed) {
+              this.ns.preset('deleted_email_success');
 
-        this._$refresh.next(true);
-      } else {
-        this.ns.preset('deleted_email_failure');
-      }
-    });
+              this._$refresh.next(true);
+            } else {
+              this.ns.preset('deleted_email_failure');
+            }
+          });
+        }
+      });
   }
 }
